moved the api/article tests to there own file for easier readability
